WIP: Move jwt table def to public
