+++ src/node/rpc/jwt_management.h
@@ -3,12 +3,9 @@
 #pragma once
 
 #include "ccf/crypto/verifier.h"
-#include "ccf/ds/json.h"
-#include "ccf/service/map.h"
-#include "ccf/service/tables/proposals.h"
+#include "ccf/service/tables/jwt.h"
 
 #include <openenclave/attestation/verifier.h>
-#include <optional>
 #include <set>
 #include <sstream>
 #if defined(INSIDE_ENCLAVE) && !defined(VIRTUAL_ENCLAVE)
@@ -19,95 +16,6 @@
 
 namespace ccf
 {
-  struct JwtIssuerKeyPolicy
-  {
-    /** OE claim name -> hex-encoded claim value
-        See openenclave/attestation/verifier.h */
-    std::optional<std::map<std::string, std::string>> sgx_claims;
-
-    bool operator!=(const JwtIssuerKeyPolicy& rhs) const
-    {
-      return rhs.sgx_claims != sgx_claims;
-    }
-  };
-
-  DECLARE_JSON_TYPE(JwtIssuerKeyPolicy);
-  DECLARE_JSON_REQUIRED_FIELDS(JwtIssuerKeyPolicy, sgx_claims);
-
-  enum class JwtIssuerKeyFilter
-  {
-    All,
-    SGX
-  };
-
-  DECLARE_JSON_ENUM(
-    JwtIssuerKeyFilter,
-    {{JwtIssuerKeyFilter::All, "all"}, {JwtIssuerKeyFilter::SGX, "sgx"}});
-
-  struct JwtIssuerMetadata
-  {
-    /// JWT issuer key filter
-    JwtIssuerKeyFilter key_filter;
-    /// Optional Key Policy
-    std::optional<JwtIssuerKeyPolicy> key_policy;
-    /// Optional CA bundle name used for authentication when auto-refreshing
-    std::optional<std::string> ca_cert_bundle_name;
-    /// Whether to auto-refresh keys from the issuer
-    bool auto_refresh = false;
-  };
-
-  DECLARE_JSON_TYPE_WITH_OPTIONAL_FIELDS(JwtIssuerMetadata);
-  DECLARE_JSON_REQUIRED_FIELDS(JwtIssuerMetadata, key_filter);
-  DECLARE_JSON_OPTIONAL_FIELDS(
-    JwtIssuerMetadata, key_policy, ca_cert_bundle_name, auto_refresh);
-
-  using JwtIssuer = std::string;
-  using JwtKeyId = std::string;
-  using Cert = std::vector<uint8_t>;
-
-  using JwtIssuers = ServiceMap<JwtIssuer, JwtIssuerMetadata>;
-  using JwtPublicSigningKeys = kv::RawCopySerialisedMap<JwtKeyId, Cert>;
-  using JwtPublicSigningKeyIssuer =
-    kv::RawCopySerialisedMap<JwtKeyId, JwtIssuer>;
-
-  namespace Tables
-  {
-    static constexpr auto JWT_ISSUERS = "public:ccf.gov.jwt.issuers";
-    static constexpr auto JWT_PUBLIC_SIGNING_KEYS =
-      "public:ccf.gov.jwt.public_signing_keys";
-    static constexpr auto JWT_PUBLIC_SIGNING_KEY_ISSUER =
-      "public:ccf.gov.jwt.public_signing_key_issuer";
-  }
-
-  struct JsonWebKey
-  {
-    std::vector<std::string> x5c;
-    std::string kid;
-    std::string kty;
-
-    bool operator==(const JsonWebKey& rhs) const
-    {
-      return x5c == rhs.x5c && kid == rhs.kid && kty == rhs.kty;
-    }
-  };
-  DECLARE_JSON_TYPE(JsonWebKey)
-  DECLARE_JSON_REQUIRED_FIELDS(JsonWebKey, x5c, kid, kty)
-
-  struct JsonWebKeySet
-  {
-    std::vector<JsonWebKey> keys;
-
-    bool operator!=(const JsonWebKeySet& rhs) const
-    {
-      return keys != rhs.keys;
-    }
-  };
-  DECLARE_JSON_TYPE(JsonWebKeySet)
-  DECLARE_JSON_REQUIRED_FIELDS(JsonWebKeySet, keys)
-
-// Unused in all sample apps, but used by node frontend
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wunused-function"
   static void remove_jwt_public_signing_keys(kv::Tx& tx, std::string issuer)
   {
     auto keys = tx.rw<JwtPublicSigningKeys>(Tables::JWT_PUBLIC_SIGNING_KEYS);
@@ -284,5 +192,4 @@ namespace ccf
 
     return true;
   }
-#pragma clang diagnostic pop
 }
