+++ src/node/rpc/node_frontend.h
@@ -16,6 +16,7 @@
 #include "frontend.h"
 #include "node/network_state.h"
 #include "node/quote.h"
+#include "node/rpc/jwt_management.h"
 #include "node/rpc/serialization.h"
 #include "node/session_metrics.h"
 #include "node_interface.h"
